app/requirements.txt
streamlit==1.28.0
langchain==0.0.335
sentence-transformers==2.2.2
elasticsearch==8.10.0
chromadb==0.4.24
pandas==2.1.1
python-dotenv==1.0.0
tqdm==4.66.1
huggingface-hub==0.12.1
requests>=2.31.0
torch==1.13.1
